requirements.txt
openstacksdk
# 5.8+ replaces tenant_id support which train still relies on
# for various "side-effects" such as creating "default" security
# groups....don't ask...
python-openstackclient == 5.7.0
tabulate
requests
python-dateutil
jinja2
dataclasses; python_version < '3.7'

# for tests
pytest
